services/content-worker/requirements.txt
# Core async + AMQP
aio-pika==9.5.5
asyncpg==0.30.0
httpx==0.28.1
python-dotenv==1.0.1

# Anthropic SDK — content generation
anthropic==0.52.0

# YouTube transcript extraction
yt-dlp==2025.3.31
youtube-transcript-api==1.0.3

# URL scraping fallback (Firecrawl not available / no API key)
requests==2.32.3
beautifulsoup4==4.13.4
lxml==5.3.2

# SSRF guard: resolve hostnames before allowing HTTP requests
dnspython==2.7.0

# Observability
structlog==25.2.0
opentelemetry-api==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-exporter-otlp-proto-grpc==1.34.0

# DB access
sqlalchemy[asyncio]==2.0.41

# Redis — rate limiter for YouTube ingestion
redis==5.2.1

# Encryption — AES-256-GCM for Hashnode API key decrypt
cryptography==44.0.2

# Pydantic for config and message schemas
pydantic==2.11.5
pydantic-settings==2.9.1

# Testing
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.1.0
respx==0.21.1
